uyren.ai_chatbot_final-main/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
supabase>=2.20.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
numpy>=1.24.0
faiss-cpu>=1.7.0
requests>=2.31.0
